requirements.txt
google-genai>=1.0
PyQt6>=6.6
python-docx>=1.0
PyMuPDF>=1.24
pint>=0.24
pydantic>=2.0
charset-normalizer>=3.0
